+++ app/src/routes/auth/login/+layout.svelte
@@ -1,5 +1,16 @@
 <script lang="ts">
+	import { goto } from '$app/navigation';
+	import { page } from '$app/stores';
 	import { Video } from '$lib/ui/base';
+	import { onMount } from 'svelte';
+
+	onMount(async () => {
+		const response = await fetch('/me');
+		if (response.ok) {
+			const redirectPage = $page.url.searchParams.get('redirect') ?? '/';
+			goto(redirectPage, { replaceState: false });
+		}
+	});
 
 	const videoOptions = {
 		sources: [
