+++ app/src/routes/auth/login/+page.svelte
@@ -1,32 +1,18 @@
 <script lang="ts">
+	import { enhance } from '$app/forms';
 	import { Button, Input } from '$lib/ui/base';
-	import { post } from '$lib/utils/post';
-	import type { RequestStep } from '$lib/types';
+	import type { ActionData } from './$types';
 
-	let request: RequestStep = 'start';
-	let username = '';
-	let errorMessage = '';
-	let magicLink = '';
+	export let form: ActionData;
 
-	async function handleSubmit() {
-		request = 'loading';
-		magicLink = '';
-		try {
-			const { accessUrl } = await post<{ accessUrl?: string }>('/auth/login', { username });
-			magicLink = accessUrl;
-			request = 'success';
-		} catch (error) {
-			request = 'error';
-			console.error(error.statusText, error.status, error.message);
-			if (error.message === 'ACCOUNT_NOT_FOUND') {
-				errorMessage = "Ce courriel n'est pas rattaché à un compte existant";
-			} else if (error.message === 'ACCOUNT_NOT_CONFIRMED') {
-				errorMessage =
-					"Ce compte n'est pas confirmé. Contactez votre administrateur de territoire de Carnet de Bord.";
-			} else {
-				errorMessage = 'Une erreur est survenue, veuillez ré-essayer ultérieurement.';
-			}
-		}
+	let loading = false;
+
+	function submitHandler() {
+		loading = true;
+		return async ({ update }) => {
+			await update();
+			loading = false;
+		};
 	}
 </script>
 
@@ -38,22 +24,23 @@
 	<div>Se connecter</div>
 	<div>au Carnet de bord</div>
 </h1>
-{#if request !== 'success'}
-	<form class="flex w-full flex-col space-y-6" on:submit|preventDefault={handleSubmit}>
+{#if !form?.success}
+	<form class="flex w-full flex-col space-y-6" method="POST" use:enhance={submitHandler}>
 		<div>Veuillez saisir votre adresse de courriel pour recevoir votre lien de connexion.</div>
 		<div class="flex w-full flex-col gap-8">
 			<div class="flex w-full flex-col gap-4">
 				<Input
 					class="!mb-0"
-					bind:value={username}
 					inputLabel="Courriel"
 					placeholder="nour@social.gouv.fr"
-					error={request === 'error' ? errorMessage : ''}
-					required={true}
+					name="username"
+					error={form?.message ? form?.message : ''}
+					required
+					value={form?.username ?? ''}
 				/>
 				<div>
-					<Button type="submit" disabled={!username || request === 'loading'}>
-						{#if request === 'loading'}
+					<Button type="submit" disabled={loading}>
+						{#if loading}
 							<span class="ri ri-refresh-line" /> Connexion...
 						{:else}
 							Se connecter
@@ -67,8 +54,8 @@
 	<div class="flex w-full flex-col gap-16">
 		<div class="flex w-full flex-col gap-6">
 			<div>Un lien vous a été envoyé pour vous connecter au Carnet de bord.</div>
-			{#if magicLink}
-				<div><a href={magicLink}>Ouvrir Carnet de bord</a></div>
+			{#if form?.accessUrl}
+				<div><a href={form?.accessUrl}>Ouvrir Carnet de bord</a></div>
 			{/if}
 		</div>
 		<div class="flex flex-col gap-6">
